+++ feature/home/src/main/res/drawable/home_bottom_navigation_item_state.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- チェック状態 -->
+    <item android:color="?colorAccent" android:state_checked="true" />
+    <!-- タップしたとき -->
+    <item android:color="?colorPrimaryDark" android:state_pressed="true" />
+    <!-- 通常 -->
+    <item android:color="#9A9EAB" />
+</selector>
