+++ feature/home/src/main/res/menu/home_bottom_navigation_menu.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/nav_home"
+        android:icon="@drawable/home_ic_home_24dp"
+        android:title="@string/common_home" />
+    <item
+        android:id="@+id/nav_library"
+        android:icon="@drawable/home_ic_video_library_24dp"
+        android:title="@string/common_library" />
+</menu>
