+++ feature/home/src/main/res/navigation/home_nav.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/navigation"
+    app:startDestination="@id/nav_home">
+
+    <fragment
+        android:id="@+id/nav_home"
+        android:name="com.github.gericass.world_heritage_client.home.HomeFragment"
+        tools:layout="@layout/home_fragment_home" />
+
+    <fragment
+        android:id="@+id/nav_library"
+        android:name="com.github.gericass.world_heritage_client.library.LibraryFragment"
+        tools:layout="@layout/library_fragment_library" />
+
+</navigation>
