+++ feature/home/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
     <string name="home_search">Search</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
