+++ feature/library/build.gradle
@@ -16,6 +16,8 @@ android {
         consumerProguardFiles 'consumer-rules.pro'
     }
 
+    resourcePrefix "library_"
+
     buildTypes {
         release {
             minifyEnabled false
@@ -30,6 +32,7 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.core:core-ktx:1.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
