+++ feature/library/src/main/java/com/github/gericass/world_heritage_client/library/LibraryFragment.kt
@@ -1,0 +1,24 @@
+package com.github.gericass.world_heritage_client.library
+
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+/**
+ * A simple [Fragment] subclass.
+ */
+class LibraryFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.library_fragment_library, container, false)
+    }
+
+
+}
