+++ feature/search/src/main/res/layout/search_fragment_search.xml
@@ -94,7 +94,7 @@
             android:id="@+id/border"
             android:layout_width="0dp"
             android:layout_height="1dp"
-            android:background="@color/common_gray_light"
+            android:background="@color/common_color_border"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/app_bar" />
